use -n for nothreads.

+++ sbin/iscsid/iscsid_main.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: iscsid_main.c,v 1.4 2012/05/27 19:52:51 christos Exp $	*/
+/*	$NetBSD: iscsid_main.c,v 1.5 2012/05/27 20:05:04 christos Exp $	*/
 
 /*-
  * Copyright (c) 2005,2006,2011 The NetBSD Foundation, Inc.
@@ -45,10 +45,8 @@
 
 list_head_t list[NUM_DAEMON_LISTS];	/* the lists this daemon keeps */
 
-#ifndef ISCSI_NOTHREAD
 pthread_mutex_t sesslist_lock;	/* session list lock */
 pthread_t event_thread;			/* event thread handle */
-#endif
 
 int driver = -1;				/* the driver's file desc */
 int client_sock;				/* the client communication socket */
@@ -57,6 +55,7 @@ int client_sock;				/* the client communication socket */
 #define ISCSI_DEBUG 0
 #endif
 int debug_level = ISCSI_DEBUG;	/* How much info to display */
+int nothreads;
 
 /*
    To avoid memory fragmentation (and speed things up a bit), we use the
@@ -170,21 +169,18 @@ init_daemon(void)
 		list[i].num_entries = 0;
 	}
 
-#ifndef ISCSI_NOTHREAD
-	if ((i = pthread_mutex_init(&sesslist_lock, NULL)) != 0) {
+	if (!nothreads && (i = pthread_mutex_init(&sesslist_lock, NULL)) != 0) {
 		printf("Mutex init failed (%d)\n", i);
 		close(sock);
 		return -1;
 	}
-#endif
 
 	if (!register_event_handler()) {
 		printf("Couldn't register event handler\n");
 		close(sock);
 		unlink(ISCSID_SOCK_NAME);
-#ifndef ISCSI_NOTHREAD
-		pthread_mutex_destroy(&sesslist_lock);
-#endif
+		if (!nothreads)
+			pthread_mutex_destroy(&sesslist_lock);
 		return -1;
 	}
 
@@ -481,9 +477,9 @@ process_message(iscsid_request_t *req, iscsid_response_t **prsp, int *prsp_temp)
 void
 exit_daemon(void)
 {
-#ifndef ISCSI_NOTHREAD
-	LOCK_SESSIONS;
-#endif
+	if (nothreads) {
+		LOCK_SESSIONS;
+	}
 	deregister_event_handler();
 
 #ifndef ISCSI_MINIMAL
@@ -516,9 +512,7 @@ main(int argc, char **argv)
 	socklen_t fromlen;
 	iscsid_request_t *req;
 	iscsid_response_t *rsp;
-#ifdef ISCSI_NOTHREAD
 	struct timeval seltout = { 2, 0 };	/* 2 second poll interval */
-#endif
 
 	client_sock = init_daemon();
 	if (client_sock < 0)
@@ -526,8 +520,11 @@ main(int argc, char **argv)
 
 	printf("iSCSI Daemon loaded\n");
 
-	while ((c = getopt(argc, argv, "d")) != -1)
+	while ((c = getopt(argc, argv, "dn")) != -1)
 		switch (c) {
+		case 'n':
+			nothreads++;
+			break;
 		case 'd':
 			debug_level++;
 			break;
@@ -538,19 +535,20 @@ main(int argc, char **argv)
 	if (!debug_level)
 		daemon(0, 1);
 
-#ifndef ISCSI_NOTHREAD
-	ret = pthread_create(&event_thread, NULL, event_handler, NULL);
-	if (ret) {
-		printf("Thread creation failed (%zd)\n", ret);
-		close(client_sock);
-		unlink(ISCSID_SOCK_NAME);
-		deregister_event_handler();
-		pthread_mutex_destroy(&sesslist_lock);
-		return -1;
+	if (nothreads)
+		setsockopt(client_sock, SOL_SOCKET, SO_RCVTIMEO, &seltout,
+		    sizeof(seltout));
+	else {
+		ret = pthread_create(&event_thread, NULL, event_handler, NULL);
+		if (ret) {
+			printf("Thread creation failed (%zd)\n", ret);
+			close(client_sock);
+			unlink(ISCSID_SOCK_NAME);
+			deregister_event_handler();
+			pthread_mutex_destroy(&sesslist_lock);
+			return -1;
+		}
 	}
-#else
-	setsockopt(client_sock, SOL_SOCKET, SO_RCVTIMEO, &seltout, sizeof(seltout));
-#endif
 
     /* ---------------------------------------------------------------------- */
 
@@ -560,17 +558,19 @@ main(int argc, char **argv)
 		fromlen = sizeof(from);
 		len = sizeof(iscsid_request_t);
 
-#ifdef ISCSI_NOTHREAD
-		do {
-			ret = recvfrom(client_sock, req, len, MSG_PEEK | MSG_WAITALL,
-							(struct sockaddr *) &from, &fromlen);
-			if (ret == -1)
-				event_handler(NULL);
-		} while (ret == -1 && errno == EAGAIN);
-#else
-		ret = recvfrom(client_sock, req, len, MSG_PEEK | MSG_WAITALL,
-					 (struct sockaddr *)(void *)&from, &fromlen);
-#endif
+		if (nothreads) {
+			ret = recvfrom(client_sock, req, len, MSG_PEEK |
+			    MSG_WAITALL, (struct sockaddr *)(void *)&from,
+			    &fromlen);
+		} else {
+			do {
+				ret = recvfrom(client_sock, req, len, MSG_PEEK |
+				    MSG_WAITALL, (struct sockaddr *) &from,
+				    &fromlen);
+				if (ret == -1)
+					event_handler(NULL);
+			} while (ret == -1 && errno == EAGAIN);
+		}
 
 		if ((size_t)ret != len) {
 			perror("Receiving from socket");
